address comments add profiling move nccl to its own file
