+++ .vale/styles/config/vocabularies/General/accept.txt
@@ -27,3 +27,5 @@ Pythonic
 [Gg]rafana
 CLI
 Anyscale
+Nsight
+[Pp]rofiler
