+++ doc/source/ray-core/doc_code/cgraph_quickstart.py
@@ -10,7 +10,7 @@ class SimpleActor:
 
 # __simple_actor_end__
 
-# __simple_actor_usage_start__
+# __ray_core_usage_start__
 import time
 
 a = SimpleActor.remote()
@@ -25,7 +25,7 @@ msg_ref = a.echo.remote("hello")
 ray.get(msg_ref)
 end = time.perf_counter()
 print(f"Execution takes {(end - start) * 1000 * 1000} us")
-# __simple_actor_usage_end__
+# __ray_core_usage_end__
 
 # __dag_usage_start__
 import ray.dag
@@ -69,7 +69,7 @@ print(f"Execution takes {(end - start) * 1000 * 1000} us")
 dag.teardown()
 # __teardown_end__
 
-# __cgraph_dependency_start__
+# __cgraph_bind_start__
 a = SimpleActor.remote()
 b = SimpleActor.remote()
 
@@ -81,7 +81,7 @@ with ray.dag.InputNode() as inp:
 
 dag = dag.experimental_compile()
 print(ray.get(dag.execute("hello")))
-# __cgraph_dependency_end__
+# __cgraph_bind_end__
 
 # __cgraph_multi_output_start__
 import ray.dag
@@ -134,46 +134,3 @@ except ray.exceptions.ActorDiedError:
 # Optionally, use the live actors to create a new Compiled Graph.
 assert live_actors == actors[1:]
 # __cgraph_actor_death_end__
-
-# __cgraph_gpu_to_gpu_start__
-import torch
-from ray.experimental.channel.torch_tensor_type import TorchTensorType
-
-ray.init()
-# Note that the following example requires at least 2 GPUs.
-assert (
-    ray.available_resources().get("GPU") >= 2
-), "At least 2 GPUs are required to run this example."
-
-
-@ray.remote(num_gpus=1)
-class GPUSender:
-    def send(self, shape):
-        return torch.zeros(shape, device="cuda")
-
-
-@ray.remote(num_gpus=1)
-class GPUReceiver:
-    def recv(self, tensor: torch.Tensor):
-        assert tensor.device.type == "cuda"
-        return tensor.shape
-
-
-sender = GPUSender.remote()
-receiver = GPUReceiver.remote()
-# __cgraph_gpu_to_gpu_end__
-
-# __cgraph_nccl_start__
-with ray.dag.InputNode() as inp:
-    dag = sender.send.bind(inp)
-    # Add a type hint that the return value of `send` should use NCCL.
-    dag = dag.with_tensor_transport("nccl")
-    # NOTE: With ray<2.42, use `with_type_hint()` instead.
-    # dag = dag.with_type_hint(TorchTensorType(transport="nccl"))
-    dag = receiver.recv.bind(dag)
-
-# Compile API prepares the NCCL communicator across all workers and schedule operations
-# accordingly.
-dag = dag.experimental_compile()
-assert ray.get(dag.execute((10,))) == (10,)
-# __cgraph_nccl_end__
